+++ src/ui/Frame/Solver.cpp
@@ -360,7 +360,7 @@ namespace two
 		{
 			//size_t column0 = frame.d_frame->dindex(d_depth);
 			size_t column = frame.d_frame->d_widget.m_index;
-			frame.d_frame->d_index[d_depth] = column;
+			frame.d_frame->d_index[d_depth] = uint(column);
 			return 1 + column < m_solvers.size() ? *m_solvers[1 + column] : *this;
 		}
 		return *this;
